Power Ups
Added a teleport shunt power up, and a double jump power up. both last for ten seconds, there is a speed one but its not working and I think its a little cringe and basic. Also Aidan, the teleoprting mihgt mess with some collision physics so dont be surprised if when you teleoprt in the direction of a car it shatters.